VillanDiffusion/my_requirements.txt
accelerate==0.19.0
comet-ml
matplotlib
datasets
tqdm
tensorboard
tensorboardX
tensorflow-datasets
einops
pytorch-fid
joblib
PyYAML
kaggle
wandb
torchsummary
torchinfo
torchmetrics
piq
lpips
